+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add support for links of all item kinds (enum variant, macro attribute, etc.)
 - Don't print extraneous newline
+- Print stderr of cargo on error even if `--quiet-cargo` is set
 
 ## [0.2.1] - 2025-07-10